+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 5.1.1 (Next)
 
+* [#57](https://github.com/mongoid/mongoid_orderable/pull/57): Ensure compatibility with mongoid 7 - [@tomasc](https://github.com/tomasc).
 * [#52](https://github.com/mongoid/mongoid_orderable/pull/52): Test against Mongoid 6 - [@dblock](https://github.com/dblock).
 * Your contribution here.